(CODEX) Extract API app factory
What changed:
- move FastAPI setup into bt_servant_engine.apps.api with a create_app() factory and explicit routers
- introduce state/dependencies modules plus health/admin/webhook routers with supporting tests
- update existing tests to hook into the new modules and add unit coverage for helpers

Why:
- begin the architectural refactor by aligning runtime layout with the planned package structure
- keep bt_servant.py as a thin shim while eliminating cross-module cycles and enforcing layering

Test plan:
- LOG_PSEUDONYM_SECRET=test-secret ENABLE_ADMIN_AUTH=false scripts/check_repo.sh
