+++ bt_servant_engine/apps/api/__init__.py
@@ -1,1 +1,5 @@
-"""API delivery subpackage placeholder."""
+"""FastAPI app factory and delivery layer exports."""
+
+from .app import create_app
+
+__all__ = ["create_app"]
