+++ bt_servant_engine/apps/api/routes/__init__.py
@@ -1,1 +1,5 @@
-"""API route module namespace for upcoming FastAPI routers."""
+"""Router namespace exports for FastAPI include hooks."""
+
+from . import admin, health, webhooks
+
+__all__ = ["admin", "health", "webhooks"]
